requirements.txt
transformers>=4.38.0
accelerate>=0.27.0
sentence-transformers
faiss-cpu
datasets
pandas
scipy
bitsandbytes
protobuf
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
nltk
tqdm
pydantic
typing-extensions
